feat: User delete exceptions handling

+++ src/main/java/com/github/VictorAlencar00/web_services/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.github.VictorAlencar00.web_services.entities.User;
 import com.github.VictorAlencar00.web_services.repositories.UserRepository;
+import com.github.VictorAlencar00.web_services.services.exceptions.DatabaseException;
 import com.github.VictorAlencar00.web_services.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +33,17 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id)) {
+				System.out.println("Error: user not found!");
+				throw new ResourceNotFoundException(id);
+			}
+			repository.deleteById(id);
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 
 	public User update(Long id, User obj) {
